Fixes and improvements to Shop and Menu testsuites
