--- src/menu/v1beta0/retrieve.js
@@ -1,152 +1,0 @@
-
-/*
- * Copyright 2018, Bloombox, LLC. All rights reserved.
- *
- * Source and object computer code contained herein is the private intellectual
- * property of Bloombox, a California Limited Liability Corporation. Use of this
- * code in source form requires permission in writing before use or the
- * assembly, distribution, or publishing of derivative works, for commercial
- * purposes or any other purpose, from a duly authorized officer of Momentum
- * Ideas Co.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- * Bloombox: Menu Retrieve
- *
- * @fileoverview Provides the ability to fetch menu data via the Menu API.
- */
-
-/*global goog */
-
-goog.provide('bloombox.menu.MenuRetrieveException');
-goog.provide('bloombox.menu.retrieveLegacy');
-
-goog.require('bloombox.config.active');
-
-goog.require('bloombox.logging.error');
-goog.require('bloombox.logging.info');
-goog.require('bloombox.logging.log');
-goog.require('bloombox.logging.warn');
-goog.require('bloombox.menu.Routine');
-goog.require('bloombox.menu.rpc.MenuRPC');
-
-
-// -- Structures -- //
-
-/**
- * Callback function type declaration for menu data retrieval.
- *
- * @typedef {function(?Object, ?number)}
- */
-bloombox.menu.MenuRetrieveCallback;
-
-
-/**
- * Represents an exception that occurred while retrieving menu data for a given
- * partner/location pair.
- *
- * @param {string} message Exception error message.
- * @constructor
- * @export
- */
-bloombox.menu.MenuRetrieveException = function MenuRetrieveException(message) {
-  /**
-   * Exception message.
-   *
-   * @type {string}
-   */
-  this.message = message;
-};
-
-
-/**
- * Public accessor method to retrieve this exception's error message.
- *
- * @return {string}
- * @export
- */
-bloombox.menu.MenuRetrieveException.prototype.getMessage = function() {
-  return this.message;
-};
-
-
-/**
- * Retrieve menu data for the current partner/location pair.
- *
- * @public
- * @param {?bloombox.menu.RetrieveOptions=} options Configuration options for
- *        this menu retrieval operation. See type docs for more info.
- * @param {bloombox.menu.MenuRetrieveCallback} callback Operation callback.
- * @throws {bloombox.menu.MenuRetrieveException} If partner/location isn't set,
- *         or some other client-side error occurs.
- * @return {Promise<proto.bloombox.services.menu.v1beta1.GetMenu.Response>}
- *         Promise attached to the underlying RPC call.
- */
-bloombox.menu.retrieveLegacy = function(options, callback) {
-  // load partner and location codes
-  let config = bloombox.config.active();
-  let partnerCode = config.partner;
-  let locationCode = config.location;
-  if (options.scope) {
-    const scopeSplit = options.scope.split('/');
-    if (scopeSplit.length !== 4)
-      throw new bloombox.menu.MenuRetrieveException(
-        'Failed to parse override scope value.');
-    partnerCode = scopeSplit[1];
-    locationCode = scopeSplit[3];
-  }
-
-  if (!partnerCode ||
-    !(typeof partnerCode === 'string' && partnerCode.length > 1) ||
-    !(typeof locationCode === 'string' && locationCode.length > 1))
-    throw new bloombox.menu.MenuRetrieveException(
-      'Partner and location must be set via `bloombox.menu.setup` before' +
-      ' retrieving menu data.');
-
-  bloombox.logging.info('Retrieving menu for \'' +
-    partnerCode + ':' + locationCode + '\'...');
-
-  return new Promise(function(resolve, reject) {
-    // stand up an RPC object
-    const rpc = new bloombox.menu.rpc.MenuRPC(
-      /** @type {bloombox.menu.Routine} */ (bloombox.menu.Routine.RETRIEVE),
-      'GET', [
-        'partners', partnerCode,
-        'locations', locationCode,
-        'global:retrieve'].join('/'));
-
-    let done = false;
-    rpc.send(function(response) {
-      if (done) return;
-      if (response !== null) {
-        done = true;
-
-        bloombox.logging.log('Received response for menu data RPC.', response);
-        if (typeof response === 'object') {
-          // dispatch the callback
-          if (callback) callback(response, null);
-          resolve(response);
-        } else {
-          bloombox.logging.error(
-            'Received unrecognized response for menu data RPC.', response);
-          if (callback) callback(null, null);
-          reject(response);
-        }
-      }
-    }, function(status) {
-      bloombox.logging.error(
-        'An error occurred while querying menu data. Status code: \'' +
-        status + '\'.');
-
-      // pass null to indicate an error
-      if (callback) callback(null, status || null);
-      reject(status);
-    });
-  });
-};
